addAssoc doesn't remove any buffer options
So you can search again for another option not included in the first
search results. Before this, after searching and adding an item on a
has_many assoc you were only able to add items that were included in the
results of that first search.

To avoid adding the same item twice, currently selected items are
excluded from the results of the search.